Create useActiveWorkspace hook function.
